movement core
