+++ declaration.d.ts
@@ -1,6 +1,3 @@
 declare module '*.scss';
 declare module '*.mp3';
-declare module '*.svg' {
-  const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
-  export default content;
-}
+declare module '*.svg';
