+++ src/core/movement/Moving/Moving.module.scss
@@ -1,0 +1,8 @@
+.movingCore {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
