+++ src/core/movement/Moving/components/BotRow/index.ts
@@ -1,0 +1,1 @@
+export { BotRow } from './BotRow';
