+++ src/core/movement/Moving/components/MidRow/index.ts
@@ -1,0 +1,1 @@
+export { MidRow } from './MidRow';
