+++ src/core/movement/Moving/components/TopRow/index.ts
@@ -1,0 +1,1 @@
+export { TopRow } from './TopRow';
