+++ src/core/movement/Moving/components/index.ts
@@ -1,0 +1,3 @@
+export { TopRow } from './TopRow';
+export { MidRow } from './MidRow';
+export { BotRow } from './BotRow';
