+++ src/core/movement/Moving/utils/index.ts
@@ -1,0 +1,10 @@
+export { chunk, directionObj, currentPlayer, currentMouse } from './constants';
+export {
+  diagonalSpd,
+  roundNumber,
+  getSpeedComponents,
+  setPlayerMove,
+  setPlayerPosition,
+  setMouseMapPosition,
+  setMousePosition,
+} from './helpers';
