+++ src/core/movement/move-btns/move-btns.tsx
@@ -1,36 +1,37 @@
+import { FC } from 'react';
+
 import { createContext, useState } from 'react';
-import { showVector, moveKeys } from './utils';
+import { moveKeys } from './utils';
 import { keyVector } from 'src/core/types/movementCore';
 
 const MoveContext = createContext<keyVector>([]);
 
-export function MoveBtns(props) {
+interface props {
+  children: React.ReactNode;
+}
+
+const MoveBtns: FC<props> = (props) => {
   const [moveKeysStack, setMoveKeysStack] = useState<keyVector>([]);
 
-  function handleMoveKeuDown(event) {
-    showVector(moveKeysStack);
+  function handleMoveKeyDown(event) {
     if (!moveKeysStack.includes(event.code) && moveKeys.includes(event.code)) {
       setMoveKeysStack([...moveKeysStack, event.code]);
     }
   }
 
-  function handleMoveKeuUp(event) {
-    showVector(moveKeysStack);
+  function handleMoveKeyUp(event) {
     if (moveKeys.includes(event.code)) {
       setMoveKeysStack([...moveKeysStack.filter((key) => key !== event.code)]);
     }
   }
 
   return (
-    <div
-      onKeyDown={(e) => handleMoveKeuDown(e)}
-      onKeyUp={(e) => handleMoveKeuUp(e)}
-    >
+    <div tabIndex={0} onKeyDown={handleMoveKeyDown} onKeyUp={handleMoveKeyUp}>
       <MoveContext.Provider value={moveKeysStack}>
         {props.children}
       </MoveContext.Provider>
     </div>
   );
-}
+};
 
-export { MoveContext };
+export { MoveContext, MoveBtns };
