+++ src/core/types/movementCore/index.ts
@@ -1,1 +1,7 @@
-export { keyVector } from './types';
+export {
+  keyVector,
+  directionVector,
+  playerState,
+  playerTest,
+  mouseTest,
+} from './types';
