+++ src/core/types/movementCore/types.ts
@@ -2,4 +2,37 @@ type keyWASD = 'KeyW' | 'KeyA' | 'KeyS' | 'KeyD';
 
 type keyVector = keyWASD[];
 
-export { keyVector };
+type directionVector = '➚' | '⬉' | '🠕' | '⬊' | '⬋' | '🠔' | '🠖' | '🠗' | 'STOP';
+type playerState = 'Стоит' | 'Движется';
+
+interface playerTest {
+  width: number;
+  height: number;
+  xPositionPx: number;
+  yPositionPx: number;
+  direction: directionVector;
+  baseSpeedPxPerSecond: number;
+  state: playerState;
+  baseSpeedChunkPerSecond: number;
+  speedPxPerSecond: number;
+  speedChunkPerSecond: number;
+  xSpeedPxPerSecond: number;
+  ySpeedPxPerSecond: number;
+  xSpeedChunkPerSecond: number;
+  ySpeedChunkPerSecond: number;
+  xPositionChunk: number;
+  yPositionChunk: number;
+}
+
+interface mouseTest {
+  screenXPositionPx: number;
+  screenYPositionPx: number;
+  screenXPositionChunk: number;
+  screenYPositionChunk: number;
+  mapXPositionPx: number;
+  mapYPositionPx: number;
+  mapXPositionChunk: number;
+  mapYPositionChunk: number;
+}
+
+export { keyVector, directionVector, playerState, playerTest, mouseTest };
